weekend update

+++ frameworks/aspnet-minimal-aot/Models.cs
@@ -7,7 +7,7 @@ sealed class DatasetItem
     public int Id { get; set; }
     public string Name { get; set; } = "";
     public string Category { get; set; } = "";
-    public double Price { get; set; }
+    public int Price { get; set; }
     public int Quantity { get; set; }
     public bool Active { get; set; }
     public List<string> Tags { get; set; } = [];
@@ -19,12 +19,12 @@ sealed class ProcessedItem
     public int Id { get; set; }
     public string Name { get; set; } = "";
     public string Category { get; set; } = "";
-    public double Price { get; set; }
+    public int Price { get; set; }
     public int Quantity { get; set; }
     public bool Active { get; set; }
     public List<string> Tags { get; set; } = [];
     public RatingInfo Rating { get; set; } = new();
-    public double Total { get; set; }
+    public long Total { get; set; }
 }
 
 sealed class DbItemDto
@@ -32,7 +32,7 @@ sealed class DbItemDto
     public int Id { get; set; }
     public string Name { get; set; } = "";
     public string Category { get; set; } = "";
-    public double Price { get; set; }
+    public int Price { get; set; }
     public int Quantity { get; set; }
     public bool Active { get; set; }
     public List<string> Tags { get; set; } = [];
@@ -41,6 +41,6 @@ sealed class DbItemDto
 
 sealed class RatingInfo
 {
-    public double Score { get; set; }
+    public int Score { get; set; }
     public int Count { get; set; }
 }
