+++ frameworks/genhttp/Tests/AsyncDatabase.cs
@@ -36,8 +36,8 @@ public class AsyncDatabase
         await using var cmd = PgDataSource.CreateCommand(
             "SELECT id, name, category, price, quantity, active, tags, rating_score, rating_count FROM items WHERE price BETWEEN $1 AND $2 LIMIT 50");
         
-        cmd.Parameters.AddWithValue((double)min);
-        cmd.Parameters.AddWithValue((double)max);
+        cmd.Parameters.AddWithValue(min);
+        cmd.Parameters.AddWithValue(max);
         
         await using var reader = await cmd.ExecuteReaderAsync();
 
@@ -50,11 +50,11 @@ public class AsyncDatabase
                 id = reader.GetInt32(0),
                 name = reader.GetString(1),
                 category = reader.GetString(2),
-                price = reader.GetDouble(3),
+                price = reader.GetInt32(3),
                 quantity = reader.GetInt32(4),
                 active = reader.GetBoolean(5),
                 tags = JsonSerializer.Deserialize<List<string>>(reader.GetString(6)),
-                rating = new { score = reader.GetDouble(7), count = reader.GetInt32(8) },
+                rating = new { score = reader.GetInt32(7), count = reader.GetInt32(8) },
             });
         }
